requirements.txt
Django>=3.2,<4.0
mysqlclient
Pillow
django-taggit==3.1.0
